requirements.txt
sqlalchemy
PyQt5
bcrypt
pyjwt
werkzeug